Fix FIPS detection
The original FIPS detection code fails on python3:

    $ python3
    Python 3.7.6 (default, Dec 19 2019, 22:52:49)
    >>> '0' == b'0'
    False

This is because bytes and strings are not directly comparable in all
scenarios, so the comparison now returns false. Python3's subprocess
also returns bytes in most scenarios:

> By default, this function will return the data as encoded bytes. The
> actual encoding of the output data may depend on the command being
> invoked, so the decoding to text will often need to be handled at the
> application level.

This results in PKI incorrectly believing that it is in FIPS mode,
when it really isn't.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>