Fix interactive installation for subsystems other than CA
When doing an interactive installation, the pkispawn script tries
to connect to Security domain via `sd_connect` and attaches user
credentials. At this point the user has not been prompted for any
credentials. So, the authentication happens with empty strings. As
a result the interactive installation fails.

This was not observed in non-interactive installation because all the info
is provided via cfg file and is available in the dictionary at the time
of execution.

This patch moves the authentication logic from `sd_connect()`
to `sd_login()` (ie) authenticate before trying to log in

The bug was introduced in commit: 08ea62892a894553d8ceae200618c6fa8d7f0585

Resolves: BZ#1795215

Signed-off-by: Dinesh Prasanth M K <dmoluguw@redhat.com>