Fixed missing token name in serverCertNick.conf
The serverCertNick.conf is used to store the nickname and
the token name of the SSL server certificate.

Previously in HSM cases the token name was missing from this
file due to mishandling, causing the installation to fail.

The SystemCertDataFactory.create() has been modified to pass
the token name properly. Also the configuration.py has been
modified to normalize the token name and use the default token
name if it's not available before storing it into the file.

https://bugzilla.redhat.com/show_bug.cgi?id=1806840