getPasswordStore(): extract to IPasswordStore
The ACME LDAP database driver needs an IPasswordStore.  But we don't
want to depend on CMSEngine, where getPasswordStore() is defined.

So extract getPasswordStore() to a static method on the
IPasswordStore interface.  (Static methods on interfaces are
supported since Java 8).  This requires one small change: cmsutil
does not depend on on cmscore so the config store must be passed as
a Map<String, String> instead of a PropConfigStore.