ACMEEngine: extract order checks to method
ACMEEngine.validateOrder() throws an exception if something is wrong
with the order (e.g. order does not match account ID, or order is
expired).  But it is useful to have a variation that does not throw
an exception and just returns the check result as a value.

Extract this logic to the 'checkOrder()' method.