Remove MAX_NUM_POLICIES hardcoded limit on cerpolicy extension
(This patch is a forward port of #349)

There are cases where an user will like to add more than 20 policies to a cert. RFC 5280
does not impose any hard limit on the number of policies. This patch removes
the hardcoded limit from the code. The user can now specify unlimited policies

This patch also fixes a minor bug to the method signature: `createUserNotice()` (ie)
the noticeNums and noticeText params have been interchanged based on the usage
in other parts of the code.

Resolves: BZ#1768718

Signed-off-by: Dinesh Prasanth M K <dmoluguw@redhat.com>