Patch ECAdminCertProfile upgrade script
Backport of PR #355

The caECAdminCert profile was added 2 years ago but was never patched
to be added to the CS.cfg. Hence, when a user tries to upgrade, the path
did not exist and so, the upgrade failed. This patch adds the missing
attribute to ensure smooth upgradation process

Move 10.8.2 upgrade script to 10.8.3

Resolves: BZ#1814242
Upstream: https://pagure.io/dogtagpki/issue/3168

Signed-off-by: Dinesh Prasanth M K <dmoluguw@redhat.com>