acme: add config to disable entire service
Read the (optional) `engine.conf' to enable or disable the entire
ACME engine.  If the file is missing or if its 'enabled=' knob is
not present, the default is to enable the service (preserving
existing behaviour).

The status is enforced by the ACMERequestFilter which will result in
status 503 Service Unavailable if the engine is disabled.