acme: add dynamic enable/disable of ACME service
Update ACME configuration system to support dynamically
enabling/disabling the ACME service.

This patch implements two providers: one that just supplies default
configuration (ACME service is uncondtionally enabled) and another
that reads the configuration from a file, and also watches that file
and reloads the configuration when the file is modified.

The engine configuration source and its parameters are defined in
'configsources.conf'.  When using ACMEEngineConfigFileSource the
file to read is configurable; 'engine.conf' is suggested.

Other parts of the ACME service configuration may eventually be made
configurable, e.g. validator configuration, backend configuration,
etc.  When we implement those features, 'configsources.conf' should
be used to configure those config sources too.

The FreeIPA implementation of ACME sercice will require LDAP-based
dynamic reconfiguration (i.e. so the ACME service can be
enabled/disabled deployment-wide with a single modification in
replicated LDAP database).  This patch does not implement an LDAP
configuration provider (that will come in a later patch).  But it is
a proof of the interfaces and may be useful for other ACME
deployments.  Some other parts of the configuration, e.g. which
challenges are enabled, may also require a similar capability so
this patch also serves as a reference implementation of the concept.