acme: reject CSR with unknown indentifiers
Unknown identifiers must be treated as unauthorised.  Otherwise a
CSR with only authorised DNS names but e.g. *also* an RFC822 name,
would be passed through to the CA and (under a standard profile
configuration) issued with the unrecognised name.

This is an unacceptable security flaw and also violates RFC 8555
§7.4.