Add symlink to apache-commons-net jar
Dependency to apache-commons-net was introduced in the commit:
https://github.com/dogtagpki/pki/commit/fca6d89dcd2b9e6592879c85a2f2278ed1a28e2f
The symlink to JARs need to be created for new instance installations as well as
existing instances.

This patch adds symlink to the new dependency.

TODO: The upgrade script to add symlinks to existing instances. This effort
should be coupled when CT moves out of prototype phase.

Signed-off-by: Dinesh Prasanth M K <dmoluguw@redhat.com>