acme: remove redundant schema file
ACME LDAP schema has been extracted as a modify LDIF.  I tested the
FreeIPA schema update machinery and it works fine with a modify
LDIF.  So the other schema LDIF, which is not an update object but a
plain entry, can be removed.

We could do likewise for LWCA and profile schema, but that is for
another day.