Added OpenLDAP database for ACME
A new OpenLDAPDatabase class has been added to provide an ACME
database in OpenLDAP. The class is currently identical to the
existing LDAPDatabase class, but it's provided for implementing
OpenLDAP- specific code in the future.

A new schema file, a sample initialization file, and a sample
configuration file for OpenLDAP have been added.

The pki-server acme-database-* commands have also been
modified to support an "openldap" type.

The ACME database documentation has been updated as well.