Handle SCEP messages sent in POST body
If the SCEP request is a HTTP POST request, we try to read the message as a binary string from the request body as outlined in https://tools.ietf.org/html/draft-gutmann-scep-16#section-4.3.

Note that the content type is ignored here - all POST request are processed this way which may cause problems when parameters are sent as form values in the body.