CI: Use NFS instead of virtualbox guest
VBGuest addition to a VM requires a rebuild of plugins, causing
a whole lot of packages to be installed. This patch removes the
dependency on VBGuest and uses NFS to sync the folders between
guest and host.

Note that since we plan to add more tests, we are cutting out
F31 from this patch (ie) we will be running our tests on latest
stable fedora release

Signed-off-by: Dinesh Prasanth M K <dmoluguw@redhat.com>