Add AppService
The AppService has been added to provide a list of subsystems
deployed on the instance. The list is generated by checking each
subfolder under <instance>/conf to see if it has a corresponding
deployed web application. The /pki/ui page has been modified to
use the AppService to provide links to all subsystems on the
instance.