startup notification: add pki-systemd jar
Implement SystemdStartupNotifier, which does not actually notify
systemd yet (this will be implemented in a subsequent commit).
Ship this class in its own jar.  The inclusion of this jar in the
pki-server package will be made conditional on an RPM macro in the
next commit.

Part of: https://pagure.io/dogtagpki/issue/1233