startup notification: make the systemd class optional
Add the `sdnotify` RPM bcond to make the SystemdStartupNotifier
class optional.  When enabled, it is supplied in a separate JAR
(part of the pki-server package), and the webapp symlink is added in
the pki-ca package.  It is enabled by default.

Note that on Fedora and presumably RHEL also, libsystemd.so is
always present.  sd_booted(3) can be used to determine whether pid 1
is systemd or not, so having a systemd notifier implementation
present doesn't imply that systemd must be used.

Nevertheless, it was requested to make this component optional.  So
here we are.

Part of: https://pagure.io/dogtagpki/issue/1233