startup notification: complete systemd notifier
Complete the implementation of SystemdStartupNotifier.  We use JNA
to bind to libsystemd.  The dependency on 'jna' package only occurs
when %{with sdnotify}.

The systemd unit template file is left alone, retaining Type=simple.
In order to enable systemd startup notification, you can override
the Type in the "drop-in" directory. For example, if the instance
name is 'pki-tomcat', write to the file:

  /etc/systemd/system/pki-tomcatd@pki-tomcat.service.d/notify.conf

the content:

  [Service]
  Type=notify

See systemd.unit(5) for more details.

Fixes: https://pagure.io/dogtagpki/issue/1233