startup notification: return result to caller
Update the StartupNotifier interface to return a NotifyResult to the
caller.  The NotifyResult contains a success/failure enum and a
string message.  Update CMSEngine to interpret the NotifyResult and
log a message if appropriate.

The SystemdStartupNotifier no longer writes to stderr/stdout.  As a
result, there is no longer any ambiguity about which CMS subsystem a
failure occurred in.

Part of: https://pagure.io/dogtagpki/issue/1233