acme: refactor validation retry behaviour into ACMEChallengeProcessor
Challenge retry behaviour was duplicated across validators.
Refactor this behaviour into the ACMEChallengeProcessor class.

The validateChallenge method can now no longer raise (arbitrary)
exceptions.  Instead it returns a value of the ValidationResult
type.  As a consequence, more exceptions need to be handled, but the
upshot is that such failures are now captured and get represented
properly in the challenge object's error field.