Add LDAPConfigMonitor
The LDAPConfigMonitor has been added for LDAPDatabase
to monitor ACME config properties stored in the database
using persistent search.

The ACMEEnableService and ACMEDisableService have been
modified to update the 'enabled' config property in the
database.

The ACMERequestFilter has been modified to use the `enabled`
config property from the database. However, if the property
is not set in the database, the filter will use the property
from the config file instead.