acme: validate identifier syntax
Update ACMENewOrderService to validate identifier syntax and respond
with an error object if it finds a malformed identifier.

Also add a service base class providing the "respond with ACMEError
object".  Other service classes should be updated over time to use
it, so that we (a) return proper error objects in conformance with
RFC 8555 and (b) avoid code duplication through use of the base
class.