CommonNameToSANDefault: allow label starting with digit
https://tools.ietf.org/html/rfc1123#section-2 relaxes the grammar
specified in https://tools.ietf.org/html/rfc1034#section-3.5,
allowing a DNS label to start with a number.  RFC 5280 explicitly
adopts the modifications of RFC 1123, so the current check that
requires a label to start with a letter is too strict.  Update the
check to allow labels to start with number or letter.

Fixes: https://github.com/dogtagpki/pki/issues/3339