Update pki pkcs12-cert-mod
The pki pkcs12-cert-mod has been modified to search
for the cert to modify in a PKCS #12 file by its ID
in addition to its nickname. If a cert ID is provided,
there will be at most one cert matching the ID. If a
nickname is provided, there could be multiple certs
matching the nickname, but only the first one will
be processed.